Added a lever interactable, added testing map
Changed the default thirdpersoncharacter blueprint to have an interact button "E" on the keyboard.

Added a blueprint interface system for interactables so that we can easily add any interactable objects within the environment (buttons, levers, doors, etc)

Created the first interactable, a lever, which as of now for testing just turns the lever gold when the player presses "E" on it.

Also added a new level for testing.

NOTE: Need to work on the player to make him 2.5D (Remove camera movement with mouse and constrain movement to the 2D plane)