Added flappy bird ability ("Q")
Added a skeleton for the flappy bird ability, probably will need to reformat where we place our abilities when we add more.

The current ability is activated for 3 seconds when the player presses "Q", allowing for the player to jump an infinite number of times for those 3 seconds at a slowed speed.

Will need to balance this ability, but this is just the first iteration and a skeleton code for how our player can interact with abilities