Added door-lever system to open doors
Player can walk up to a lever and interact with it using "E". This lever will open and close a door, if a door is assigned to the lever. Otherwise, the lever will just turn on and off.

To assign a door to a lever, after creating a lever instance of a blueprint inside a level and a door instance, you have to go to the details panel of the lever and find  the drop down menu for the DoorReference, assigning the correct door there.