Made player head flat
Potentially temporary fix for making the player's head flat on top to avoid being flung away from objects when jumping and hitting the corner of your head on them.

Unreal Engine 5 seems to not allow for editing of the normal capsule mesh for collisions unless you make your own movement systems, so this temporary fix works around it by adding a small box collision on top of the capsule with the same radius as the capsule to ensure all head hits are flat on surfaces.

Note that this may be changed in the future as this can lead to more issues, but for now it works in place of the prior movement.