added button
BP_Button completed, subject to revision. Button activates for a set amount of time determined by the level editor setting.

When placing a button in a level, you have the option to set a float value on the button pertaining to the amount of time that the button should stay activated based on when the user presses it.