Added sound settings to main menu
Sound settings added to main menu, but it is just a base implementation and doesn't save any values when closing the game or even closing the UI widget.

The sound sliders do work, however, and currently only effect the music on the main menu. Will need to complete the implementation with integration inside the level, changing all sound assets to match the classes provided by the sound mixers. Also need to add a saving mechanism for this feature, so that when you swap between levels the sound is saved correctly.