Level balancing and fixes
Balanced the level to make the gameplay a bit easier in terms of button timing. Also changed one of the moving platform sections a bit to be more intuitive.

Fixed the settings menu inconsistency with the main menu level, now it should work correctly there as well without any issues.

Changed the volume slightly on the music cue, and made the main menu play the music using the cue instead of the asset itself.