Added hazard to reset player to a location
BP_Hazard is a box. When the player collides with this box, which can be shaped within the level editor, he is reset to a location that can be defined within the level editor by selecting the hazard component in the level. This location by default is (0, 0, 0), which is the origin of the level.

Added a test of this hazard to the SideScrollerTesting level and refactored a bit of code in the BP_Fountain blueprint.