Changed flappy bird ability to be activated through fountain
Flappy bird ability is now toggleable with the use of the fountain. We can use this or something else as charge stations for the ability.

Also made some minor changes to the ability and to the fountain grayscale effect.

The flappy bird ability now only slows the player after the first in-air "flap" meaning that the player has normal movement until attempting a second jump.

The fountain can now be set in the level to start the level as grayscale or not, which we can experiment with whether we want the grayscale effect in the game.