Small fixes
Found a few small bugs from the playtesting day, including the cases for levers being too small for their new models and interactable objects being activated by non-player actors such as doors or moving platforms.

Ensured that interactable objects are only interacted with if overlapping with a player character, and increased the size of the cases to accommodate for the lever size change.