Added map selection level
Added a level called MapSelect that functions to select which level you are going to play.

Added blueprint for level selector which is an actor and can be interacted with, selecting and opening a level.

Also edited the BP_PlayerCharacter to contain a new camera for the MapSelect level, which is an aerial view.