Added breakable platforms
Added breakable platform option to the BP_MovingPlatform blueprint. Checking this box and setting a timer on the instance of a MovingPlatform blueprint will cause the platform to break after a set duration when the player touches it.

Fixed/edited the BP_Lever material to set back to the original material when deactivated.