New Music assets
Level 1 music assets:
Intro chord for the very beggining of level
layer 1 for when bloop isn't moving.
layer 2 for bloop moving. This layers should implemented on top of layer 1 with a short fade in
layer 3: implement this layer on top of layer 1 and 2 after x amount of time
If bloop stays still for more than 5 seconds we start the layer process above.
Ability layer 1 : implement when ability is active
ability layer 2: implement on top of ability 1 on for variety if needed