Implemented lever animation
Removed the old lever's on/off logic changing the material of the lever. New lever uses the animation to turn on and off.

Note: the turning off animation doesn't work. Not sure how to reverse an animation in unreal, but the reverse is implemented and it simply snaps the lever to the off position as of now.

Also fixed the breakable logic in BP_MovingPlatform