Added button animation, fixed lever animation
Added button animation and skeletal mesh to replace the old button.

Found out how to reverse an animation. Just needed to set the position of the active animation playing to be at the end before reversing the play rate (with a value of -1 play rate).

Fixed the lever animation to be consistent forwards and backwards with animations.

Changed the lever material to be a skeletal mesh material rather than a static mesh material, avoiding issues with exports.