added parachute ability
BP_Parachute is an actor component blueprint that holds all the logic for the parachuting ability.

To test the ability, it is currently on a toggleable button "Q" to activate and deactivate the parachute.

I separated the logic so that we don't need to directly change BP_PlayerCharacter whenever we make a change to this ability.

The ability is set up as components of the BP_PlayerCharacter blueprint and calls an event DeployParachute from the BP_Parachute actor component.