Fixed animations a bit
Animation graph now just skips the jump start animation sequence because I couldn't get it to blend well with the rest of the jump.

There are pros and cons to this, but overall the new jump cycle looks a lot better

Still a noticeable issue when trying to loop jumps (holding down spacebar and jumping right when you hit the ground).