Started working on a rope swing mechanic
Very first iteration of a rope swing mechanic.

Definitely not optimal, but the current version shoots a rope straight up from bloop and attaches itself to a ceiling.

When bloop attaches to the ceiling, he has a constant force applied towards that attach point until he detaches, which is similar to a swinging mechanic, but still very barebones and inaccurate.