Added RopeSwing ability to BP_PlayerCharacter
Added rope swing as an actor component for BP_PlayerCharacter, separating the rope swing logic from the rest of the character logic.

Also removed the invisible walls on level 2, as it currently breaks the ability. We don't need invisible walls once we restrict the player to planar movement.

The only reason the planar movement isn't pushed is because it's easier to test physics in a 3D environment rather than a 2D one, but for our main build it should definitely be constrained.