More work on rope swinging
Should work now as intended, but there are major issues with multiple ropes being set up near each other.

Need to address how interactions work with the player, as if the player interacts with another grapple point while swinging on a different rope, it breaks the physics.

We need a way to make sure the player cannot interact with other grapple points while he is on a rope, and also a way to interact with the grapple point that is closest to him.