Started working on frog pull points
Prototyped version of an actor that can be placed in the level as a point for the player to use the frog tongue to pull themselves towards a point.

Using launch character for now because it was easy to prototype, will likely change to a linear pull or something a bit different.