Added instance-editability for grapple hook interaction ranges
It is now possible to edit how far you can interact with a grapple hook from the level editor.

Uses construction script which runs every time an actor is created or edited to dynamically set the size of the interaction boxes to the instances defined values set within the level editor.