Added first water wheel implementation (doesn't move player)
Added BP_WaterWheel, which dynamically creates instances of platforms that rotate like a water wheel, allowing the player to platform on them.

Instance editability includes a full water wheel rotation time, the size of the platforms, and the amount of platforms on the water wheel.

Probably have to rework some of this logic because the player doesn't follow the platform movement velocity right now.