Fixed some issues with wind and grapple visuals
Wind no longer has interaction box collision showing in-game

Made a material for grapple points for when they are able to be interacted with by the player

Reworked cable component logic on various blueprints: BP_PlayerCharacter, BP_FrogPullPoint, and BP_RopeSwing to instantiate a new cable each time then destroy it when the cable is not being used. This fixes a visual issue where the one cable being used was stretching is awkward angles when being attached to new grapple points.