Slightly fixed bp_hazard camera reset
BP_Hazard is collecting a duplicate and incorrect transform value from BP_PlayerCharacter's camera component for the charater view.

I added a Do Once node with a delay for a reset on it to make sure only one value is passed through each time.

There are still inconsistencies, but it should be a lot more rare for issues to occur