Added music layering to level 1
First implementation of music layering. Only uses idle, walk, and intro soundtracks for now.

Implementation is an actor blueprint that is placed in the level called BP_MusicManager that starts by running the intro soundtrack, the moves on to the walking soundtrack. When the player idles for 5 seconds, the soundtrack fades into the idle state, and when the player starts moving again fades back to the walking soundtrack.