Added functionality to deactivate tutorial triggers
Tutorial triggers can now be deactivated so that when something in the game happens such as opening a certain door, that tutorial trigger will no longer be activatable.

This mitigates the previous undesired effect where the tutorial text would appear telling you to open a door that is already opened.