Fixed level 2 reference for build
There were two level script actors being generated for this one level, leading to an error when building the game with this level in it.

Not sure how they got generated, but the only way I was able to remove it was by using an online solution:

Make the level blueprint abstract, reparent it to itself, then make it not abstract anymore and reparent again.