Music Level 2
Hello guys!
Long time no see. I'm sending the stems for level 2 music here.

Here's a breakdown of the files:
SECTION 1. That's an example of not moving vs moving. So you'll hear the track start with no drums. That is how it would be if there's no movement. Then the drums start and that would be for movement. So the easiest implementation would be to take all the layers and use the drums in and out as a condition for movement.

Then there's SECTION 2, which I think would be great for ability moment. This track has 3 parts so it doesn't get boring.

Then there's SECTION 3 which is very similar to section 1 but is thought out to come after section 2 so is not exactly the same as before the ability.

But use the layers as you like, it doesn't have to be like I mention above. :)

Right now the layers are printed from beggining to end (Sec 1, 2 and 3 all toghether) So they just need to be lined up with each other. I am happy to print the layers in sections if you think is needed.

And last I want to say hi to everyone and send you a big hug!! I'm really sorry I have not been at the meetings. Life is very busy and I barely have no time between projects. I told Zara that I'm happy to have a music specific meeting anytime if you want and we can make deadlines for Level 3 music or anything else.

Talk soon guys!!!