SFX Ambience and Assets for Level 01
Made an ambient sfx blueprint that triggers when overlapped with collision box near first log.  Contains birds, wind gusts, branch shakes, insects, and an lfo-controlled wind Metasounds asset.   There is also a metasounds file that is triggered containing leaves and a shimmery pad-like sound (meant to evoke the particles in the air and the magic of the forest).  I realize this is a bit of a musical element, but I think it fits nicely with everything.  Once the blueprint is triggered, all of these elements continuously sound at random intervals of time and distance from the character location.

Also made an actor blueprint containing a sound cue with a variety of grass sounds that can be triggered at random places upon the walking path.